+++ src/plugin/babel.spec.js
@@ -1,6 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
 import { isEntryFile } from "./babel";
+import { createContent } from "./html";
 
 const readFixture = (filename) =>
   fs.readFileSync(path.join(__dirname, "../../fixtures", filename), "utf-8");
@@ -32,5 +33,13 @@ describe("isEntryFile", () => {
     const src = readFixture(filename);
     const paths = isEntryFile(src, filename);
     expect(paths).toEqual(false);
+  });
+});
+
+describe("createContent", () => {
+  it("export default webViewRender(App)", () => {
+    const filename = "app-export-default.jsx";
+    const src = readFixture(filename);
+    expect(createContent(src)).toMatchSnapshot();
   });
 });
